extracted logic csv-logic from staticsController
